+++ src/components/AppElements/components/CodeEditor/hooks/emailEditorHooks.js
@@ -34,6 +34,31 @@ const emailEditorHooks = {
 		});
 	},
 
+	onEditSubjectLineAndPreviewText: ({ subjectLine, previewText }) => {
+		return new Promise(resolve => {
+			resolve();
+		});
+	},
+
+
+	onAiAssistant: ({ context, action, tone, customContext,	target }) => {
+		return new Promise(resolve => {
+			resolve({ result });
+		});
+	},
+
+	onExternalElementDropIn: ({ externalElementId, elementJson, defaultJson }) => {
+		return new Promise(resolve => {
+			resolve(elementJson);
+		});
+	},
+
+	onExternalElementButtonClicked: ({ externalElementId, buttonId, elementJson, defaultJson }) => {
+		return new Promise(resolve => {
+			resolve(elementJson);
+		});
+	},
+
 	onEditImage: async ({ originalImage, lockDimensions }) => {
 		await galleryInstance.methods.updateData({ currentImgSrc: originalImage, dimensions: lockDimensions });
 		await galleryInstance.show();