+++ src/components/EmailEditor/components/Settings.vue
@@ -232,7 +232,46 @@
 								color="primary"
 								hide-details
 								:true-value="true"
-								@change="updateActionMenu($event, name)"
+								@change="updateBActionMenu($event, name)"
+							/>
+						</v-col>
+					</v-row>
+				</v-card>
+			</v-card>
+		</OptionWrapper>
+
+		<h3>Element Action Menu</h3>
+		<p>
+			You can control what your users capable to do with elements. Drag, edit, duplicate or delete. All this functionality can be turned off.
+		</p>
+		<OptionWrapper>
+			<v-card elevation="0" class="d-flex px-2">
+				<v-card
+					min-height="72px"
+					flat
+					class="rounded-0 rounded-t d-flex pa-4"
+					width="100%"
+				>
+					<v-row>
+						<v-col
+							v-for="(value, name) in elementActions"
+							:key="name"
+							cols="6"
+							class="d-flex justify-space-between align-center"
+						>
+							<v-card-title
+								class="ma-0 pa-0 text-subtitle-1"
+								style="margin-bottom: -3px !important"
+							>
+								{{ name }}
+							</v-card-title>
+							<v-switch
+								:input-value="elementActions[name]"
+								class="ma-0 pa-0"
+								color="primary"
+								hide-details
+								:true-value="true"
+								@change="updateEActionMenu($event, name)"
 							/>
 						</v-col>
 					</v-row>
@@ -456,6 +495,45 @@
 				</v-card>
 			</v-card>
 		</OptionWrapper>
+
+		<h3>Subject Line and Preview Text</h3>
+		<p>
+			You can toggle the editing of the subject line and preview text as well
+		</p>
+		<OptionWrapper>
+			<v-card elevation="0" class="d-flex px-2">
+				<v-card
+					min-height="72px"
+					flat
+					class="rounded-0 rounded-t d-flex pa-4"
+					width="100%"
+				>
+					<v-row>
+						<v-col
+							v-for="(value, name) in subjectLineAndPreviewText"
+							:key="name"
+							cols="6"
+							class="d-flex justify-space-between align-center"
+						>
+							<v-card-title
+								class="ma-0 pa-0 text-subtitle-1"
+								style="margin-bottom: -3px !important"
+							>
+								{{ name }}
+							</v-card-title>
+							<v-switch
+								:input-value="subjectLineAndPreviewText[name]"
+								class="ma-0 pa-0"
+								color="primary"
+								hide-details
+								:true-value="true"
+								@change="updateSubjectLineAndPreviewTextOption($event, name)"
+							/>
+						</v-col>
+					</v-row>
+				</v-card>
+			</v-card>
+		</OptionWrapper>
 	</div>
 </template>
 
@@ -519,6 +597,9 @@ export default {
 		blockActions() {
 			return this.$store.state.editorConfig.settings.actionMenu.block;
 		},
+		elementActions() {
+			return this.$store.state.editorConfig.settings.actionMenu.element;
+		},
 		dropZones() {
 			return this.$store.state.editorConfig.settings.dropzones;
 		},
@@ -543,6 +624,9 @@ export default {
 		title() {
 			return this.$store.state.editorConfig.settings.title;
 		},
+		subjectLineAndPreviewText() {
+			return this.$store.state.editorConfig.settings.subjectLineAndPreviewText;
+		},
 	},
 	methods: {
 		processName(name) {
@@ -554,8 +638,12 @@ export default {
 			this.updateDropZones({ [name]: !!value });
 			this.updateEditorSettings();
 		},
-		updateActionMenu(value, name) {
+		updateBActionMenu(value, name) {
 			this.updateBlockActionMenu({ [name]: !!value });
+			this.updateEditorSettings();
+		},
+		updateEActionMenu(value, name) {
+			this.updateElementActionMenu({ [name]: !!value });
 			this.updateEditorSettings();
 		},
 		updateToolbox(value, name) {
@@ -578,11 +666,16 @@ export default {
 			this.updateTitle({ [name]: !!value });
 			this.updateEditorSettings();
 		},
+		updateSubjectLineAndPreviewTextOption(value, name) {
+			this.updateSubjectLineAndPreviewText({ [name]: !!value });
+			this.updateEditorSettings();
+		},
 		...mapMutations([
 			"updateUser",
 			"updateAutoSave",
 			"updateStaticAssets",
 			"updateBlockActionMenu",
+			"updateElementActionMenu",
 			"updateDropZones",
 			"updateToolboxes",
 			"updateVideoElementBaseUrl",
@@ -590,6 +683,7 @@ export default {
 			"updateComponentPermissions",
 			"updatePanels",
 			"updateTitle",
+			"updateSubjectLineAndPreviewText",
 		]),
 		...mapActions({
 			updateEditorSettings: "updateEditorSettings",